HotFix: WorldGuard non-exist doesn't stop LM now.
